+++ R/corrected_tau.R
@@ -49,8 +49,6 @@ corrected_tau <- function(data, dvar, pvar, mvar,
     is_logical(repeated)
   )
   
-  tau_method <- tau_method[1]
-
   # prepare scdf ----
   if (missing(dvar)) dvar <- dv(data) else dv(data) <- dvar
   if (missing(pvar)) pvar <- phase(data) else phase(data) <- pvar
@@ -128,13 +126,13 @@ corrected_tau <- function(data, dvar, pvar, mvar,
   
   
   out <- list(
-    tau = sapply(x, function(x) if(is.na(x$p[1]) || x$p[1] > 0.05) x$tau[2] else x$tau[3]), 
-    p = sapply(x, function(x) if(is.na(x$p[1]) || x$p[1] > 0.05) x$p[2] else x$p[3]), 
+    tau = sapply(x, function(x) if(is.na(x$p[1]) || x$p[1] > alpha) x$tau[2] else x$tau[3]), 
+    p = sapply(x, function(x) if(is.na(x$p[1]) || x$p[1] > alpha) x$p[2] else x$p[3]), 
     corrected_tau = x,
     auto_tau = sapply(x, function(x) x$tau[1]),
     tau_corrected = sapply(x, function(x) x$tau[3]),
     tau_uncorrected = sapply(x, function(x) x$tau[2]),    
-    correction = sapply(x, function(x) if(is.na(x$p[1]) || x$p[1] > 0.05) FALSE else TRUE),
+    correction = sapply(x, function(x) if(is.na(x$p[1]) || x$p[1] > alpha) FALSE else TRUE),
     alpha = alpha,
     continuity = continuity,
     repeated   = repeated,